HV: * Big thing - the 'high resolution time stamp' patch.       completely reworked how time stamps and sample rates are kept       internally: they're now boost::rational<>s. jive5ab can now       correctly deal with data rates and time stamps up to at least 128Gbps

+++ evlbi5a/mk5command/file2check.cc
@@ -68,7 +68,7 @@ string file2check_fn(bool qry, const vector<string>& args, runtime& rte ) {
             const string            filename( OPTARG(2, args) );
             const string            strictopt( OPTARG(3, args) );
             const headersearch_type dataformat(rte.trackformat(), rte.ntrack(),
-                                               (unsigned int)rte.trackbitrate(),
+                                               rte.trackbitrate(),
                                                rte.vdifframesize());
 
             // If there's no frameformat given we can't do anything
