+++ evlbi5a/mk5command/fill2out.cc
@@ -77,7 +77,7 @@ string fill2out_fn(bool qry, const vector<string>& args, runtime& rte ) {
             const string            start_s( OPTARG(2, args) );
             const string            inc_s( OPTARG(3, args) );
             const headersearch_type dataformat(rte.trackformat(), rte.ntrack(),
-                                               (unsigned int)rte.trackbitrate(),
+                                               rte.trackbitrate(),
                                                rte.vdifframesize());
 
             EZASSERT2(dataformat.valid(), cmdexception,
