+++ evlbi5a/mk5command/net2file.cc
@@ -161,7 +161,7 @@ string net2file_fn(bool qry, const vector<string>& args, runtime& rte ) {
             // Now that we have all commandline arguments parsed we may
             // construct our headersearcher
             const headersearch_type dataformat(rte.trackformat(), rte.ntrack(),
-                                               (unsigned int)rte.trackbitrate(),
+                                               rte.trackbitrate(),
                                                rte.vdifframesize());
 
             // set read/write and blocksizes based on parameters,
